Delete unnecessary code
